[refactor] LottoJudge 삭제, RewardCalculator 이름 변경
