+++ src/main/java/lotto/domain/Lottos.java
@@ -1,5 +1,8 @@
 package lotto.domain;
 
+import lotto.common.type.WinnerRank;
+import lotto.common.type.WinnerRankCondition;
+import lotto.service.Reward;
 import lotto.strategy.LottoCommonGenerateStrategy;
 
 import java.util.ArrayList;
@@ -25,5 +28,21 @@ public class Lottos {
 
     public int getLottosAmount(){
         return lottos.size();
+    }
+
+    public static final int SECOND_PLACE_CORRECT_COUNT = 5;
+
+    public Reward countWinningNumbers(WinningLotto winningLotto) {
+        Reward reward = new Reward();
+        lottos.forEach(lotto -> {
+            int sameSize = lotto.getSameElementsSize(winningLotto.getWinningLotto());
+            boolean correctBonus = getCorrectBonus(winningLotto.getBonusWinningNumber(), lotto, sameSize);
+            reward.plusCount(WinnerRank.valueOf(WinnerRankCondition.missCountFrom(sameSize, correctBonus)));
+        });
+        return reward;
+    }
+
+    private boolean getCorrectBonus(int bonusNumber, Lotto lotto, int sameSize) {
+        return sameSize == SECOND_PLACE_CORRECT_COUNT && lotto.hasSameElement(bonusNumber);
     }
 }
