+++ src/main/java/lotto/service/Reward.java
@@ -4,19 +4,18 @@ import lotto.common.type.WinnerRank;
 import lotto.domain.Lottos;
 
 import java.util.EnumMap;
-import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Objects;
 
 import static lotto.input.TicketPriceInput.DEFAULT_TICKET_PRICE;
 
-public class RewardCalculator {
+public class Reward {
 
     private final EnumMap<WinnerRank, Integer> winnerCount = new EnumMap<>(WinnerRank.class);
 
     private final static int DEFAULT_OFFSET = 1;
 
-    public RewardCalculator() {
+    public Reward() {
         winnerCount.put(WinnerRank.FIRST, 0);
         winnerCount.put(WinnerRank.SECOND, 0);
         winnerCount.put(WinnerRank.THIRD, 0);
@@ -49,7 +48,7 @@ public class RewardCalculator {
     public boolean equals(Object o) {
         if (this == o) return true;
         if (o == null || getClass() != o.getClass()) return false;
-        RewardCalculator that = (RewardCalculator) o;
+        Reward that = (Reward) o;
         return Objects.equals(winnerCount, that.winnerCount);
     }
